% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mglrm.R
\name{mglrm}
\alias{mglrm}
\title{Multigroup Latent Regression Item Response Model}
\usage{
mglrm(Y, Ymis = c("ignore", "incorrect"), X = NULL, S = NULL, itermcmc,
  burnin, thin = 1, tdf = 10, cartctrl1 = 5, cartctrl2 = 1e-04)
}
\arguments{
\item{Y}{data frame containing item responses. They can be binary or ordinal items. The
responses must be coded starting at \code{0} or as \code{NA}. Rows of \code{Y} correspond to
persons and columns correspond to items.}

\item{Ymis}{character string how to treat \code{NA} in \code{Y}. The default method will
omit them element-wise (unbalanced panel structure) and \code{incorrect} will treat them as
incorrect answers.}

\item{X}{data frame containing person-level covariates. They can be numeric or factor
variables and contain missing values coded as \code{NA}. Rows of \code{X} correspond to
persons and columns correspond to covariates. With \code{X} set to NULL (default) an
interecpt-only model will be estimated.}

\item{S}{integer vector of individual group membership. A multigroup model with \code{S}
stratifying the sample will be estimated. The most simple latent regression item response
model without grouping results from \code{S} set to \code{NULL} (default).}

\item{itermcmc}{number of MCMC iterations.}

\item{burnin}{number of burnin iterations.}

\item{thin}{thinning interval, i.e., retain only every \code{thin}th iteration (if argument
\code{thin} is used, \code{itermcmc*thin} and \code{burnin*thin} yield the total number of
MCMC and burnin iterations).}

\item{tdf}{degrees of freedom of multivariate-t proposal distribution for category cutoff
parameters for ordinal items.}

\item{cartctrl1}{minimum number of observations in any terminal CART node during covariates
imputation cycles.}

\item{cartctrl2}{complexity parameter. Any CART split that does not decrease the overall
lack of fit by a factor of \code{control2} is not attempted during covariates imputation
cycles.}
}
\value{
list with elements \code{MCMCdraws} and \code{M-Hacc} containing a list of posterior
samples matrices (rows correspond to iterations and columns to parameters) and, if
estimated, a vector of Metropolis-Hastings acceptance rates of category cutoff parameters
for ordinal items.
}
\description{
Estimate multigroup latent regression models for binary and ordinal item response data
considering partially missing covariate data.
}
\details{
\code{mglrm} uses a fully Bayesian estimation approach. Independent conjugate prior
distributions are chosen to develop a Metropolis-within-Gibbs sampling algorithm based on
the device of data augmentation (Tanner & Wong, 1987). The function generates a sample from
the posterior distribution of a one dimensional two-parameter normal ogive IRT model
(Albert, 1992) \deqn{y_{ij}^*=\alpha_j \theta_i - \beta_j + \varepsilon_{ij}} including a
(multivariate) regression equation of person level predictors on the mean vector of latent
abilities \deqn{\theta_i=x_i\gamma_{S_i} + e_i.} Regression parameters are allowed to vary
across observed groups. Partially observed person-level covariates are imputed in each
sampling iteration. Sequential CART (Burgette & Reiter, 2010) are utilized as approximations
to the full conditional distributions of missing values in \code{X}.
}
\examples{
## prepare data input
data(simdata_2mglrm)
Y <- simdata_2mglrm[, grep("Y", names(simdata_2mglrm), value = TRUE)]
X <- simdata_2mglrm[, grep("X", names(simdata_2mglrm), value = TRUE)]

## estimation setup: MCMC chains of length 20 with 4 initial burn-in samples
## for testing purposes (for your applications itermcmc > 10000 needed)
##
## not considering grouping
results_model1 <- mglrm(Y = Y, X = X, itermcmc = 10, burnin = 2, thin = 2)

## considering grouping
results_model2 <- mglrm(Y = Y, X = X, S = simdata_2mglrm$S, itermcmc = 10, burnin = 2,
  thin = 2)
}
\references{
Albert, J. H. (1992). Bayesian estimation of normal ogive item response curves
using gibbs sampling. \emph{Journal of Educational Statistics}, \emph{17}(3), 251-269.

Burgette, L. F., & Reiter, J. P. (2010). Multiple imputation for missing data
via sequential regression trees. \emph{American Journal of Epidemiology}, \emph{172}(9),
1070-1076.

Tanner, M. A., & Wong, W. H. (1987). The calculation of posterior distributions
by data augmentation. \emph{Journal of the American Statistical Association},
\emph{82}(398), 528-549.
}
