% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rilrm.R
\name{rilrm}
\alias{rilrm}
\title{Random Intercept Latent Regression Item Response Model}
\usage{
rilrm(Y, Ymis = c("ignore", "incorrect"), X = NULL, S, itermcmc, burnin,
  thin = 1, tdf = 10, cartctrl1 = 5, cartctrl2 = 1e-04)
}
\arguments{
\item{Y}{data frame containing item responses. They can be binary or ordinal items. The
responses must be coded starting at \code{0} or as \code{NA}. Rows of \code{Y} correspond to
persons and columns correspond to items.}

\item{Ymis}{character string how to treat \code{NA} in \code{Y}. The default method will
omit them element-wise (unbalanced panel structure) and \code{incorrect} will treat them as
incorrect answers.}

\item{X}{data frame containing person-level covariates. They can be numeric or factor
variables and contain missing values coded as \code{NA}. Rows of \code{X} correspond to
persons and columns correspond to covariates. With \code{X} set to NULL (default) an
interecpt-only model will be estimated.}

\item{S}{integer vector of individual cluster membership. A random intercept model with
\code{S} stratifying the sample will be estimated.}

\item{itermcmc}{number of MCMC iterations.}

\item{burnin}{number of burnin iterations.}

\item{thin}{thinning interval, i.e., retain only every \code{thin}th iteration (if argument
\code{thin} is used, \code{itermcmc*thin} and \code{burnin*thin} yield the total number of
MCMC and burnin iterations).}

\item{tdf}{degrees of freedom of multivariate-t proposal distribution for category cutoff
parameters for ordinal items.}

\item{cartctrl1}{minimum number of observations in any terminal CART node during covariates
imputation cycles.}

\item{cartctrl2}{complexity parameter. Any CART split that does not decrease the overall
lack of fit by a factor of \code{control2} is not attempted during covariates imputation
cycles.}
}
\value{
list with elements \code{MCMCdraws} and \code{M-Hacc} containing a list of posterior
samples matrices (rows correspond to iterations and columns to parameters) and, if
estimated, a vector of Metropolis-Hastings acceptance rates of category cutoff parameters
for ordinal items.
}
\description{
Estimate random intercept latent regression models for binary and ordinal item response data
considering partially missing covariate data.
}
\details{
\code{rilrm} uses a fully Bayesian estimation approach. Independent conjugate prior
distributions are chosen to develop a Metropolis-within-Gibbs sampling algorithm based on
the device of data augmentation (Tanner & Wong, 1987). The function generates a sample from
the posterior distribution of a one dimensional two-parameter normal ogive IRT model
(Albert, 1992) \deqn{y_{cij}^*=\alpha_j \theta_{ci} - \beta_j + \varepsilon_{cij}} including
a (multivariate) regression equation of a cluster level random intercept and person level
predictors on the mean vector of latent abilities
\deqn{\theta_{ci}=\omega_c + x_{ci}\gamma + e_{ci}.} This corresponds to the most basic
multilevel specification (Fox & Glas, 2001). Partially observed person-level covariates are
imputed in each sampling iteration. Sequential CART (Burgette & Reiter, 2010) are utilized
as approximations to the full conditional distributions of missing values in \code{X}.
}
\examples{
## prepare data input
data(simdata_40rilrm)
Y <- simdata_40rilrm[, grep("Y", names(simdata_40rilrm), value = TRUE)]
X <- simdata_40rilrm[, grep("X", names(simdata_40rilrm), value = TRUE)]

## estimation setup: MCMC chains of length 20 with 4 initial burn-in samples
## for testing purposes (for your applications itermcmc > 10000 needed)
##
results <- rilrm(Y = Y, X = X, S = simdata_40rilrm$S, itermcmc = 10, burnin = 2, thin = 2)
}
\references{
Albert, J. H. (1992). Bayesian estimation of normal ogive item response curves
using gibbs sampling. \emph{Journal of Educational Statistics}, \emph{17}(3), 251-269.

Burgette, L. F., & Reiter, J. P. (2010). Multiple imputation for missing data
via sequential regression trees. \emph{American Journal of Epidemiology}, \emph{172}(9),
1070-1076.

Fox, J.-P., & Glas, C. A. W. (2001). Bayesian estimation of a multilevel irt
model using Gibbs sampling. \emph{Psychometrika}, \emph{66}(2), 271-288.

Tanner, M. A., & Wong, W. H. (1987). The calculation of posterior distributions
by data augmentation. \emph{Journal of the American Statistical Association},
\emph{82}(398), 528-549.
}
